* added honey command handler

+++ honeypot/kraken-honey-sshd/src/main/java/org/krakenapps/termio/TerminalDecoder.java
@@ -129,7 +129,7 @@ public class TerminalDecoder {
 		}
 	}
 
-	private void notifyData(int b) {
+	private void notifyData(byte b) {
 		for (TerminalEventListener listener : session.getListeners()) {
 			try {
 				listener.onData(session, b);
